build_proto(ggplot2usr)	R Documentation

builds proto widgets

Description

The function build_proto is called to build the proto widgets

Note

Unlike S3 or S4 class, the structure of proto objects is lost when the package is built embedded them in a function which will be called to create them.

Author(s)

cedric

References

http://wiener.math.csi.cuny.edu/pmg/gWidgets/index.html/ especially ProtoExample

[Package ggplot2usr version 0.1 Index]