ggplot2usr(ggplot2usr)	R Documentation

ggplot2 user interface

Description

ggplot_user_interface.R

Details

Package: 	 ggplot2usr

Type: 	 Package

Version: 	 0.1

Date: 	 2009-11-04

License: 	 GPL (>= 2)

LazyLoad: 	 yes

package is intented to be used by beginners who will be able to access to the different layers of the ggplot2 package is yet far from providing the full capability of the ggplot2, but will help in building the first graphs and print their formula : this package is still in developpement and full of bugs, so far intended only for showing the aim of the project has been inially developped within a french project, which builds a database for migratory fishes control stations, along with several graphical tools help the users to 'view' their data

TODO

include the scales, test under package.

Author(s)

Cedric Briand cedric.briand00@gmail.com

References

http://had.co.nz/ggplot2/

See Also

http://had.co.nz/ggplot2/

Examples

mtcars=mtcars # the dataframe used are listed in the base environment .GlobalEnv  
ggplot2usr(data="mtcars")

[Package ggplot2usr version 0.1 Index]